refactor(node): extract proof_history initialization job into init.rs
